requirements.txt
python-decouple==3.8
numpy==1.26.2
pandas==2.1.3
requests==2.31.0
selenium==4.16.0
webdriver-manager==4.0.1
bs4==0.0.2
